Move QuantizeScaled from Wavenet to gluonts.transform
